feat(security): implement authentication and authorization using JWT

+++ backend/api-seu-cafezinho/src/main/java/com/seucafezinho/api_seu_cafezinho/service/impl/AddressServiceImpl.java
@@ -18,6 +18,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.UUID;
 
+import static com.seucafezinho.api_seu_cafezinho.util.SecurityUtil.validateOwnership;
+
 @RequiredArgsConstructor
 @Service
 public class AddressServiceImpl implements AddressService {
@@ -28,17 +30,20 @@ public class AddressServiceImpl implements AddressService {
     @Transactional(readOnly = true)
     public AddressResponseDto findById(UUID addressId) {
         Address address = findAddressById(addressId);
+        validateOwnership(address.getUser().getId());
         return AddressMapper.INSTANCE.toDto(address);
     }
 
     @Transactional(readOnly = true)
     public Page<AddressResponseDto> findAllByUser(UUID userId, Pageable pageable) {
+        validateOwnership(userId);
         return addressRepository.findAllByUserId(userId, pageable)
                 .map(AddressMapper.INSTANCE::toDto);
     }
 
     @Transactional
     public AddressResponseDto create(AddressRequestDto createDto, UUID userId) {
+        validateOwnership(userId);
         User user = findUserById(userId);
 
         if (addressRepository.existsByUserIdAndStreetIgnoreCase(userId, createDto.getStreet())) {
@@ -55,6 +60,7 @@ public class AddressServiceImpl implements AddressService {
     @Transactional
     public AddressResponseDto update(UUID addressId, AddressRequestDto updateDto) {
         Address existingAddress = findAddressById(addressId);
+        validateOwnership(existingAddress.getUser().getId());
 
         if (!existingAddress.getStreet().equalsIgnoreCase(updateDto.getStreet()) &&
                 addressRepository.existsByUserIdAndStreetIgnoreCase(existingAddress.getUser().getId(), updateDto.getStreet())) {
@@ -71,6 +77,7 @@ public class AddressServiceImpl implements AddressService {
     @Transactional
     public void delete(UUID addressId) {
         Address address = findAddressById(addressId);
+        validateOwnership(address.getUser().getId());
         addressRepository.delete(address);
     }
