+++ backend/api-seu-cafezinho/src/main/java/com/seucafezinho/api_seu_cafezinho/service/impl/OrderServiceImpl.java
@@ -22,6 +22,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.UUID;
 
+import static com.seucafezinho.api_seu_cafezinho.util.SecurityUtil.validateOwnership;
+
 @RequiredArgsConstructor
 @Service
 public class OrderServiceImpl implements OrderService {
@@ -35,6 +37,7 @@ public class OrderServiceImpl implements OrderService {
     @Transactional(readOnly = true)
     public OrderResponseDto findById(UUID orderId) {
         Order order = findOrderById(orderId);
+        validateOwnership(order.getUser().getId());
         return OrderMapper.INSTANCE.toDto(order);
     }
 
@@ -46,12 +49,14 @@ public class OrderServiceImpl implements OrderService {
 
     @Transactional(readOnly = true)
     public Page<OrderResponseDto> findAllByUserId(UUID userId, Pageable pageable) {
+        validateOwnership(userId);
         return orderRepository.findAllByUserId(userId, pageable)
                 .map(OrderMapper.INSTANCE::toDto);
     }
 
     @Transactional
     public OrderResponseDto createOrder(UUID userId, OrderRequestDto orderRequestDto) {
+        validateOwnership(userId);
         User user = findUserById(userId);
 
         Order order = orderFactory.createOrder(user, orderRequestDto);
@@ -68,6 +73,7 @@ public class OrderServiceImpl implements OrderService {
     @Transactional
     public OrderResponseDto updateOrder(UUID orderId, OrderRequestDto orderRequestDto) {
         Order order = findOrderById(orderId);
+        validateOwnership(order.getUser().getId());
 
         orderFactory.updateOrder(order, orderRequestDto);
         order.calculateTotalPrice();
@@ -104,9 +110,11 @@ public class OrderServiceImpl implements OrderService {
     @Transactional
     public void delete(UUID orderId) {
         Order order = findOrderById(orderId);
+        validateOwnership(order.getUser().getId());
         orderRepository.delete(order);
     }
 
+    @Transactional(readOnly = true)
     private Order findOrderById(UUID orderId) {
         return orderRepository.findById(orderId)
                 .orElseThrow(() -> new EntityNotFoundException(
