+++ backend/api-seu-cafezinho/src/main/java/com/seucafezinho/api_seu_cafezinho/service/impl/UserServiceImpl.java
@@ -11,10 +11,13 @@ import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.UUID;
+
+import static com.seucafezinho.api_seu_cafezinho.util.SecurityUtil.validateOwnership;
 
 @RequiredArgsConstructor
 @Service
@@ -22,9 +25,11 @@ public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
     private final EmailProducer emailProducer;
+    private final PasswordEncoder passwordEncoder;
 
     @Transactional(readOnly = true)
     public UserResponseDto findById(UUID id) {
+        validateOwnership(id);
         User user = findUserById(id);
         return UserMapper.INSTANCE.toDto(user);
     }
@@ -41,6 +46,8 @@ public class UserServiceImpl implements UserService {
             throw new IllegalArgumentException("This email is already registered.");
         }
 
+        createDto.setPassword(passwordEncoder.encode(createDto.getPassword()));
+
         User userToSave = UserMapper.INSTANCE.toUser(createDto);
         User savedUser = userRepository.save(userToSave);
 
@@ -51,7 +58,10 @@ public class UserServiceImpl implements UserService {
 
     @Transactional
     public UserResponseDto update(UUID id, UserRequestDto updateDto) {
+        validateOwnership(id);
         User existingUser = findUserById(id);
+
+        updateDto.setPassword(passwordEncoder.encode(updateDto.getPassword()));
 
         UserMapper.INSTANCE.updateUserFromDto(updateDto, existingUser);
 
@@ -61,13 +71,20 @@ public class UserServiceImpl implements UserService {
 
     @Transactional
     public void delete(UUID id) {
+        validateOwnership(id);
         User user = findUserById(id);
         userRepository.delete(user);
     }
 
     @Transactional(readOnly = true)
-    private User findUserById(UUID id) {
+    public User findUserById(UUID id) {
         return userRepository.findById(id).orElseThrow(
                 () -> new EntityNotFoundException(String.format("User with id: '%s' not found", id)));
+    }
+
+    @Transactional(readOnly = true)
+    public User findByEmail(String email) {
+        return userRepository.findByEmail(email).orElseThrow(
+                () -> new EntityNotFoundException(String.format("User with email: '%s' not found", email)));
     }
 }
